refactor: separation of main into service
