+++ db/init/schema.sql
@@ -1,5 +1,4 @@
 CREATE TABLE `destinations` (
-  -- `id` bigint NOT NULL AUTO_INCREMENT,
   `name` varchar(100) NOT NULL,
   `description` text CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci,
   `contact_info` varchar(255) DEFAULT NULL,
